Streaming Module Update
* Streaming Module Update
* Version Bump
* Requirements update
