+++ requirements.txt
@@ -1,5 +1,6 @@
 altgraph==0.17.3
 blinker==1.6.2
+Brotli==1.1.0
 certifi==2022.12.7
 charset-normalizer==3.1.0
 click==8.1.3
@@ -21,6 +22,7 @@ m3u8==3.6.0
 MarkupSafe==2.1.2
 more-itertools==9.1.0
 mpmath==1.3.0
+mutagen==1.47.0
 networkx==3.1
 numba==0.56.4
 numpy==1.23.5
@@ -30,6 +32,7 @@ Pillow==9.3.0
 prettytable==3.6.0
 psutil==5.9.5
 PyAudio==0.2.13
+pycryptodomex==3.19.0
 pyinstaller==5.10.0
 pyinstaller-hooks-contrib==2023.2
 python-dateutil==2.8.2
@@ -37,7 +40,7 @@ pytz==2023.3
 pytz-deprecation-shim==0.1.0.post0
 pywin32-ctypes==0.2.0
 regex==2023.3.23
-requests==2.28.2
+requests==2.31.0
 semantic-version==2.10.0
 setuptools-rust==1.5.2
 six==1.16.0
@@ -46,10 +49,15 @@ SpeechRecognition==3.10.0
 sympy==1.11.1
 tiktoken==0.3.1
 tomli==2.0.1
+torch==2.1.0+cu118
+torchaudio==2.1.0+cu118
+torchvision==0.16.0+cu118
 tqdm==4.65.0
 typing_extensions==4.5.0
 tzdata==2023.3
 tzlocal==4.3
-urllib3==1.26.15
+urllib3==2.1.0
 wcwidth==0.2.6
+websockets==12.0
 Werkzeug==2.3.4
+yt-dlp==2023.11.16
